Fix issues from review
